Refactor setMiddleC (move it to it's own module) and add chord transposing
